more support for command log and more code reuse for contexts
